phoenix: Allow building custom upstreamed kernels
Snippet from google documentation -

Updating devices from Android 10

Android 10 introduces a new build flag, PRODUCT_OTA_ENFORCE_VINTF_KERNEL_REQUIREMENTS. For devices launched with Android 10, this flag is automatically set to true. When the flag is set to true, a script extracts the kernel version and the kernel configurations from the installed kernel image and writes this information to the OTA package metadata. OTA clients on devices running Android 10 or higher read this information to check compatibility.

If the script fails to extract this information for your kernel image, do one of the following:

    Set BOARD_KERNEL_VERSION to the kernel version and BOARD_KERNEL_CONFIG_FILE to the path of the built kernel configuration file .config. Both variables must be updated when the kernel image is updated.
    Alternatively, set PRODUCT_OTA_ENFORCE_VINTF_KERNEL_REQUIREMENTS to false to remove kernel requirements from the OTA package.

You can view the source code of the kernel information extraction script extract_kernel.py

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Change-Id: Ib72907d7138f5dfbe2034c12fb43049398145a79
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: CryllicBuster273 <cryllicbuster273@pixelexperience.org>